# simd draw fast paths