`nim c -d:release --gc:arc --debugger:native -r .\experiments\benchmark_cairo_draw.nim`

before:
```
pixie draw normal .................. 2.293 ms      2.396 ms    ±0.112  x1000
pixie draw overwrite ............... 1.903 ms      2.012 ms    ±0.091  x1000
pixie draw basic ................... 0.415 ms      0.432 ms    ±0.018  x1000
pixie draw mask .................... 2.251 ms      2.321 ms    ±0.071  x1000
```

with opaque optimization:
```
pixie draw normal .................. 1.052 ms      1.109 ms    ±0.041  x1000
pixie draw overwrite ............... 1.054 ms      1.138 ms    ±0.070  x1000
pixie draw basic ................... 0.460 ms      0.478 ms    ±0.017  x1000
pixie draw mask .................... 1.045 ms      1.097 ms    ±0.040  x1000
```

with this pr:
```
pixie draw normal .................. 0.617 ms      0.653 ms    ±0.039  x1000
pixie draw overwrite ............... 0.646 ms      0.671 ms    ±0.025  x1000
pixie draw basic ................... 0.441 ms      0.462 ms    ±0.021  x1000
pixie draw mask .................... 0.856 ms      0.892 ms    ±0.038  x1000
```